CurriculumMasterDto에서 curriculumId 타입을 Integer에서 Long으로 변경 및 menu_instructor.html의 메뉴 항목을 instructor 경로로 수정

+++ admin/src/main/java/kr/or/ddit/dto/CurriculumMasterDto.java
@@ -12,7 +12,7 @@ import lombok.EqualsAndHashCode;
 public class CurriculumMasterDto {
 
     @NotNull
-    private Integer curriculumId;
+    private Long curriculumId;
     @NotBlank
     private String title;
     @NotBlank
